Uploaded files including SQL_SORTER
Sql_sorter uses the JDBC library - mysql-connector-j-9.2.0.jar >>It deletes, creates, and edits entries
Driver contains the main file
Displays displays menu and options and error and warning messages
BackUPJson parses the entries into a JSON file while backuptxt stores data in a text file. 
Prompters is just for prompts and contains filter function >> filter functions meaning that if the program is asking for an integer it stops any input that is of Datatype String and character and will display an error.